generics Example
In the above program we have created a customised list which can take value of any type 
but for that 
1) declare the class name along with T
2) Declare a method with T

Generics means parameterized types. The idea is to allow type (Integer, String, … etc., and user-defined types) to be a parameter to methods, classes, and interfaces. Using Generics, it is possible to create classes that work with different data types. An entity such as class, interface, or method that operates on a parameterized type is a generic entity.